Renamed `databricks_catalog_workspace_binding` to `databricks_workspace_binding` (#3703)
* rename resource

* fix test
