+++ catalog/resource_catalog_workspace_binding_test.go
@@ -106,103 +106,6 @@ func TestCatalogWorkspaceBindingsReadOnly_Create(t *testing.T) {
 	}.ApplyNoError(t)
 }
 
-func TestSecurableWorkspaceBindings_Create(t *testing.T) {
-	qa.ResourceFixture{
-		Fixtures: []qa.HTTPFixture{
-			{
-				Method:   "PATCH",
-				Resource: "/api/2.1/unity-catalog/bindings/catalog/my_catalog",
-				ExpectedRequest: catalog.UpdateWorkspaceBindingsParameters{
-					Add: []catalog.WorkspaceBinding{
-						{
-							BindingType: catalog.WorkspaceBindingBindingTypeBindingTypeReadOnly,
-							WorkspaceId: int64(1234567890101112),
-						},
-					},
-					SecurableName: "my_catalog",
-					SecurableType: "catalog",
-				},
-				Response: catalog.WorkspaceBindingsResponse{
-					Bindings: []catalog.WorkspaceBinding{
-						{
-							BindingType: catalog.WorkspaceBindingBindingTypeBindingTypeReadOnly,
-							WorkspaceId: int64(1234567890101112),
-						},
-					},
-				},
-			}, {
-				Method:   "GET",
-				Resource: "/api/2.1/unity-catalog/bindings/catalog/my_catalog?",
-				Response: catalog.WorkspaceBindingsResponse{
-					Bindings: []catalog.WorkspaceBinding{
-						{
-							BindingType: catalog.WorkspaceBindingBindingTypeBindingTypeReadOnly,
-							WorkspaceId: int64(1234567890101112),
-						},
-					},
-				},
-			},
-		},
-		Resource: ResourceCatalogWorkspaceBinding(),
-		Create:   true,
-		HCL: `
-		securable_name = "my_catalog"
-		securable_type = "catalog"
-		workspace_id   = "1234567890101112"
-		binding_type   = "BINDING_TYPE_READ_ONLY"
-		`,
-	}.ApplyNoError(t)
-}
-
-func TestSecurableWorkspaceBindings_Delete(t *testing.T) {
-	qa.ResourceFixture{
-		Fixtures: []qa.HTTPFixture{
-			{
-				Method:   "PATCH",
-				Resource: "/api/2.1/unity-catalog/bindings/catalog/my_catalog",
-				ExpectedRequest: catalog.UpdateWorkspaceBindingsParameters{
-					Remove: []catalog.WorkspaceBinding{
-						{
-							BindingType: catalog.WorkspaceBindingBindingTypeBindingTypeReadOnly,
-							WorkspaceId: int64(1234567890101112),
-						},
-					},
-					SecurableName: "my_catalog",
-					SecurableType: "catalog",
-				},
-				Response: catalog.WorkspaceBindingsResponse{
-					Bindings: []catalog.WorkspaceBinding{
-						{
-							BindingType: catalog.WorkspaceBindingBindingTypeBindingTypeReadOnly,
-							WorkspaceId: int64(1234567890101112),
-						},
-					},
-				},
-			}, {
-				Method:   "GET",
-				Resource: "/api/2.1/unity-catalog/bindings/catalog/my_catalog?",
-				Response: catalog.WorkspaceBindingsResponse{
-					Bindings: []catalog.WorkspaceBinding{
-						{
-							BindingType: catalog.WorkspaceBindingBindingTypeBindingTypeReadOnly,
-							WorkspaceId: int64(1234567890101112),
-						},
-					},
-				},
-			},
-		},
-		Resource: ResourceCatalogWorkspaceBinding(),
-		Delete:   true,
-		ID:       "1234567890101112|catalog|my_catalog",
-		HCL: `
-		securable_name = "my_catalog"
-		securable_type = "catalog"
-		workspace_id   = "1234567890101112"
-		binding_type   = "BINDING_TYPE_READ_ONLY"
-		`,
-	}.ApplyNoError(t)
-}
-
 func TestCatalogWorkspaceBindingsReadImport(t *testing.T) {
 	qa.ResourceFixture{
 		Fixtures: []qa.HTTPFixture{
